+++ app/src/main/java/com/yrkj/yrlife/ui/LoginActivity.java
@@ -180,9 +180,9 @@ public class LoginActivity extends BaseActivity {
                             editor.putString("isBind","");
                         }
                         if (user.getTotal_balance()==null){
-                            editor.putLong("money", 0);
+                            editor.putFloat("money", 0);
                         }else {
-                            editor.putLong("money", user.getTotal_balance().longValue());
+                            editor.putFloat("money", user.getTotal_balance().floatValue());
                         }
                         editor.putString("openid", "");
                         editor.putString("access_token", "");
