+++ app/src/main/java/com/yrkj/yrlife/ui/SignActivity.java
@@ -228,9 +228,9 @@ public class SignActivity extends BaseActivity {
                                 editor.putString("isBind","");
                             }
                             if (user.getTotal_balance()==null){
-                                editor.putLong("money", 0);
+                                editor.putFloat("money", 0);
                             }else {
-                                editor.putLong("money", user.getTotal_balance().longValue());
+                                editor.putFloat("money", user.getTotal_balance().floatValue());
                             }
                             editor.putInt("jifen", user.getCard_total_point());
                             //提交修改
